changed the grammar to work more closely to the compiler parser
this makes it easier to make sure it parses the same language, and
might also help maintaining the grammar as the language changes.

one thing that's still not solved is function application because
there's some blackmagic happening in the compiler there that
I don't know how to replicate in tree-sitter.